+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Use the non-blocking dial option to connect OpenTelemetry agent via OTLP. (#VARLOG-637/#552)
+- Initialize log stream id generator without refreshing which calls RPC to storage nodes. (#VARLOG-634/#553)
 
 ### Removed
 - Remove unnecessary file `TOPIC`. (#VARLOG-631/#548)